First batch of file additions for cross-compiler:
	new file:   ghc/ghc-cross.wrapper
	new file:   includes/mkDerivedConstants.cross.awk
	new file:   includes/mkSizeMacros.cross.awk
	new file:   rules/cross-compiling.mk

These are expected to sit quietly in the tree until
the rest of the machinery matures on an (upcoming)
branch. Reviews will begin to make sense as soon as
that has happened. Anyway, comments are welcome. See
<http://www.haskell.org/pipermail/cvs-ghc/2012-July/074456.html>
for background.

Disclaimer: these source files are not (yet) up to the
            quality standards set by the rest of the tree.

Cleanups, move-arounds and rewrites (i.e. .awk -> .hs), as
well as additional comments and documentation will happen
as soon as the basic functionality of a cross-compiler is
working reliably.