Tweak the HC_OPTS generation
libraries/base/GHC/Base_HC_OPTS wasn't getting included, because we
were using libraries/base/./GHC/Base.lhs as the path, and the ./
meant we got the wrong filename.

I didn't use $(realpath ...) as the .hs file may be generated by hsc2hs
or similar, so may not exist when the HC_OPTS are constructed.