Make the call to chooseBoxingStrategy lazy again
I made it strict, as an incidental consequence of this patch:

  commit 5bae803a18b17bdb158a7780e6b6ac3c520e5b39
  Author: Simon Peyton Jones <simonpj@microsoft.com>
  Date:   Sat Sep 15 23:09:25 2012 +0100
      Fix UNPACK with -fomit-interface-pragmas.

But it's very important that chooseBoxingStrategy is lazy, else
(in bigger programs with lots of recursion in types) GHC can
loop. This showed up in Data.Sequence; and I think it was making
haddock loop as well.

Anyway this patch makes it lazy again.