Add kind-defaulting in simplifyInfer (fixes Trac #7332)
The basic point here is described in TcSimplify
   Note [Promote _and_ default when inferring]
The new thing is that, when figuring out the predicates
to abstact over in simplifyInfer, we must default OpenKind
to *, just as we do in quantifyTyVar. I had not realised
how important this was until Oleg came up with Trac #7332.

As usual I did some refactoring, so the patch affects
many more lines than strictly necessary.