Some refactoring of the occurs check in TcUnify
I was experimenting with allowing unification with a type variable
when there are type functions in the type; but that led to strangely
worse error messages (See Note [Conservative unification check]) in
TcUnify.  I don't quite understand why, but the simple thing for now
is to maintain the same behaviour.  I tidied up the code a bit though.