Make Rank2Types and PolymorphicComponents into a synonym for RankNTypes (no deprecation)
There was a long thread on cvs-ghc which concluded that although it's
the Right Thing to deprecate Rank2Types and PolymorphicComponents, in
favour of RankNTypes, it would cause lots of busy-work for library
authors and the pain isn't worth the gain.

So this patch removes the deprecation, and documents the synonym-ity.