Accurately report usage of newtype data constructors in FFI declarations
See Note [Newtype constructor usage in foreign declarations] in TcForeign.
It's quite non-trivial to say which newtype constructor are used in
foreign import/export declarations, and I had to do a bit of refactoring
to achieve it.  (Say hello to the X5 bus from Oxford to Cambridge.)
It's a bit tiresome, with some more plumbing, but not hard.

Trac #7048 triggered this change.