Add ":info!" to GHCi.  This shows all instances without filtering first.
The default behavior of :info is to show only those instances of
for a type, where all relevant type constructor names are in scope.
This keeps down the number of instances shown to the user.

In some cases, it is nice to be able to see all instances for a type.
This patch implements this with the :info! command.