Add a -rpath entry for the RTS library, so that it can find libffi
This fixes dynamic library resolution when --enable-new-dtags is used
(#7062).

When --enable-new-dtags is used when linking an executable, a RUNPATH as
well as RPATH is set. The linker then ignores RPATH, and RUNPATH is only
used for directly (not transitively) needed libraries. As the program
doesn't directly need libffi, it isn't found.