Ensure the LLVM codegen correctly handles literals in a branch. #7571
We need to be sure that when generating code for literals, we properly narrow
the type of the literal to i1. See Note [Literals and branch conditions] in the
LlvmCodeGen.CodeGen module.

This occurs rarely as the optimizer will remove conditional branches with
literals, however we can get this situation occurring with hand written Cmm
code.

This fixes Trac #7571.

Signed-off-by: David Terei <davidterei@gmail.com>