Add -f[no-]warn-unsupported-llvm-version. Closes Trac #7579.
This controls whether or not the compiler warns if we're using an LLVM
version that's too old or too new. It's mostly useful when building the
compiler knowingly with an unsupported version, so you don't get a lot
of warnings in the build process.

There's no documentation for this since it's a flag only a few
developers would care about anyway.

Signed-off-by: Austin Seipp <mad.one@gmail.com>